# internal/service/sail/docs/swagger.yaml
definitions:
...
  model.AddProject:
    properties:
      name:
        type: string
      project_group_id:
        type: integer
    type: object
...
  model.AddPublish:
    properties:
      config_id_arr:
        items:
          type: integer
        type: array
      namespace_id:
        type: integer
      project_id:
        type: integer
      publish_data:
        type: string
      publish_type:
        type: integer
    type: object
...
  model.EditProject:
    properties:
      name:
        type: string
      project_id:
        type: integer
    type: object
...
  model.ProjectList:
    properties:
      create_by:
        type: integer
      create_by_name:
        type: string
      create_time:
        type: integer
      key:
        type: string
      managed:
        type: boolean
      name:
        type: string
      project_group_id:
        type: integer
      project_group_name:
        type: string
      project_id:
        type: integer
    type: object
...
  model.PublishConfigList:
    properties:
      create_by:
        type: integer
      create_by_name:
        type: string
      create_time:
        type: integer
      namespace_id:
        type: integer
      namespace_name:
        type: string
      publish_configs:
        items:
          type: string
        type: array
      publish_id:
        type: integer
      publish_type:
        type: integer
      publish_type_str:
        type: string
      status:
        type: integer
      status_str:
        type: string
    type: object
...
  model.RollbackPublish:
    properties:
      publish_id:
        type: integer
    type: object
...
  page.Page:
    properties:
      count:
        type: integer
      list: {}
    type: object
  response.JsonResponse:
    properties:
      code:
        description: 业务码
        type: integer
      data: {}
      message:
        description: 描述信息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: FengYu
    url: https://hyy-yu.space
  description: 配置中心接口设计文档
  title: 配置中心
  version: "1.0"
paths:
...
  /v1/project/add:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.AddProject'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加项目
      tags:
      - 项目管理
  /v1/project/del:
    post:
      parameters:
      - description: ProjectId
        in: body
        name: project_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除项目
      tags:
      - 项目管理
  /v1/project/edit:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.EditProject'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 编辑项目
      tags:
      - 项目管理
  /v1/project/list:
    get:
      parameters:
      - default: 1
        description: 页号
        in: query
        name: page_index
        type: integer
      - default: 10
        description: 页长
        in: query
        name: page_size
        type: integer
      - description: 排序字段
        in: query
        name: sort
        type: string
      - description: 项目ID
        in: query
        name: project_id
        type: integer
      - description: 项目名称
        in: query
        name: project_name
        type: string
      responses:
        "200":
          description: data
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/page.Page'
                  - properties:
                      List:
                        $ref: '#/definitions/model.ProjectList'
                    type: object
              type: object
      summary: 项目列表
      tags:
      - 项目管理
...
  /v1/publish/add:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.AddPublish'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加发布
      tags:
      - 发布管理
  /v1/publish/list:
    get:
      parameters:
      - default: 1
        description: 页号
        in: query
        name: page_index
        type: integer
      - default: 10
        description: 页长
        in: query
        name: page_size
        type: integer
      - description: 排序字段
        in: query
        name: sort
        type: string
      - description: 项目ID
        in: query
        name: project_id
        required: true
        type: integer
      responses:
        "200":
          description: data
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/page.Page'
                  - properties:
                      List:
                        $ref: '#/definitions/model.PublishConfigList'
                    type: object
              type: object
      summary: 发布列表
      tags:
      - 发布管理
  /v1/publish/rollback:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.RollbackPublish'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加发布
      tags:
      - 发布管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"